git-svn-id: https://openspotlight.dev.java.net/svn/openspotlight/trunk/source/osl@180 65b4daa9-cd5d-0410-b3a2-f97e3e1d3c1c

+++ osl-graph-persistence/src/main/java/org/openspotlight/graph/persistence/SLPersistentTreeFactoryImpl.java
@@ -69,9 +69,11 @@ public class SLPersistentTreeFactoryImpl extends SLPersistentTreeFactory {
 	/* (non-Javadoc)
 	 * @see org.openspotlight.graph.persistence.SLPersistentTreeFactory#createPersistentTree()
 	 */
-	public SLPersistentTree createPersistentTree() throws SLPersistentTreeFactoryException {
+	public SLPersistentTree createTempPersistentTree(boolean removeExistent) throws SLPersistentTreeFactoryException {
 		try {
-			deleteDir(new File("/tmp/repository"));
+			if (removeExistent) {
+				deleteDir(new File("/tmp/repository"));
+			}
 			Credentials credentials = new SimpleCredentials("username", "password".toCharArray());
 			RepositoryConfig config = RepositoryConfig.create(new File(".", "src/main/resources/repository.xml").getAbsolutePath(), "/tmp/repository");
 			Repository repo = RepositoryImpl.create(config);
